requirements.txt

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.25.0

# Computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

# Machine learning utilities
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Logging and experiment tracking
wandb>=0.16.0
tensorboard>=2.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
h5py>=3.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
unittest-xml-reporting>=3.2.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Model-specific dependencies
timm>=0.9.0  # For various vision models
einops>=0.7.0  # For tensor operations

# Optional but recommended
jupyterlab>=4.0.0  # For notebooks
ipywidgets>=8.1.0  # For interactive widgets

# GPU acceleration (install based on your CUDA version)
# torch==2.1.0+cu118  # For CUDA 11.8
# torch==2.1.0+cu121  # For CUDA 12.1
